src/backend/requirements.txt
# FastAPI Backend Requirements
# ShizishanGPT Milestone 6

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/AI Libraries
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
scikit-learn==1.3.2
joblib==1.3.2

# Vector Database
chromadb==0.4.22

# Database
pymongo==4.6.0

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2

# Translation
googletrans==4.0.0rc1

# Image Processing
Pillow==10.1.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Optional (if not already installed)
numpy==1.24.3
pandas==2.1.3
